+++ src/index.ts
@@ -51,7 +51,7 @@ export { default as KeyboardShortcut } from './components/keyboard-shortcut'
 export { default as KeyCapturer, SUPPORTED_KEYS } from './components/key-capturer'
 export { default as ProgressBar } from './components/progress-bar'
 export { default as Time } from './components/time'
-export * from './components/menu'
+export * from './new-components/menu'
 
 //
 // Deprecated components